feat: initial client classes/interface
