+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+  "semi": true,
+  "useTabs": false,
+  "bracketSpacing": true
+}
